tests/wpt/tests/html/semantics/selectors/pseudo-classes/dir-html-input-dynamic-text.html

<!doctype html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1622900">
<link rel="help" href="https://html.spec.whatwg.org/#the-directionality">
<input value="ltr" dir="auto">
<script>
test(function() {
  let input = document.querySelector("input");
  assert_true(input.matches(":dir(ltr)"), "Input with ltr value should match dir(ltr)");
  input.textContent = "ﷺ";
  assert_true(input.matches(":dir(ltr)"), "Should still match dir(ltr) after text change");
  input.value = "ltr2";
  assert_true(input.matches(":dir(ltr)"), "Should still match dir(ltr) after value change");
  input.value = "ﷺ";
  assert_true(input.matches(":dir(rtl)"), "Should match dir(rtl) after value change");
  input.textContent = "ltr";
  assert_true(input.matches(":dir(rtl)"), "Should match dir(rtl) after text change");
}, ":dir on <input> isn't altered by text children")
</script>
